requirements.txt
fastapi==0.111.0
pillow==10.3.0
torch==2.3.1
ultralytics==8.2.32
uvicorn==0.30.1
